pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xwebetl"
version = "0.1.0"
description = "A flexible web content extraction, transformation, and loading (ETL) pipeline"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "ObergXData", email = "obergxdata@gmail.com"}
]
keywords = ["rss", "web scraping", "etl", "data pipeline", "content extraction"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]

dependencies = [
    "lxml>=4.9.0",
    "requests>=2.31.0",
    "feedparser>=6.0.0",
    "PyYAML>=6.0",
    "openai>=1.0.0",
    "pypdfium2>=3.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.11.0",
]

[project.urls]
Homepage = "https://github.com/obergxdata/WebETL"
Documentation = "https://github.com/obergxdata/WebETL#readme"
Repository = "https://github.com/obergxdata/WebETL"
Issues = "https://github.com/obergxdata/WebETL/issues"

[project.scripts]
webetl = "xwebetl.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["xwebetl*"]
exclude = ["test*", "data*", "test_server*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
